# [WIP] Performance improvements to Span.Slice