--- a/src/vm/mscorlib.h
+++ b/src/vm/mscorlib.h
@@ -716,6 +716,11 @@ DEFINE_METHOD(UNSAFE,               BYREF_AS,               As, GM_RefTFrom_RetR
 DEFINE_METHOD(UNSAFE,               OBJECT_AS,              As, GM_Obj_RetT)
 DEFINE_METHOD(UNSAFE,               BYREF_ADD,              Add, GM_RefT_Int_RetRefT)
 DEFINE_METHOD(UNSAFE,               BYREF_INTPTR_ADD,       Add, GM_RefT_IntPtr_RetRefT)
+#ifdef _TARGET_64BIT_
+DEFINE_METHOD(UNSAFE,               BYREF_NUINT_ADD,        Add, GM_RefT_ULong_RetRefT)
+#else
+DEFINE_METHOD(UNSAFE,               BYREF_NUINT_ADD,        Add, GM_RefT_UInt_RetRefT)
+#endif // _TARGET_64BIT_
 DEFINE_METHOD(UNSAFE,               PTR_ADD,                Add, GM_PtrVoid_Int_RetPtrVoid)
 DEFINE_METHOD(UNSAFE,               BYREF_BYTE_OFFSET,      ByteOffset, NoSig)
 DEFINE_METHOD(UNSAFE,               BYREF_ADD_BYTE_OFFSET,  AddByteOffset, NoSig)