… src/vm/metasig.h