This updates `Span<T>.Slice` and `ReadOnlySpan<T>.Slice` to zero-extend rather than sign-extend the index when computing the new base address. It's a proof of concept demonstrating the performance gains we can get from addressing https://github.com/dotnet/coreclr/issues/23666. I hacked together a prototype that performs `mov` elimination in all cases, then re-introduced this PR and reran the benchmarks to come up with the following results.

|         Method |  Toolchain |     Mean |     Error |    StdDev | Ratio |
|--------------- |----------- |---------:|----------:|----------:|------:|
| Utf8Parser_Sum | 3.0-master | 20.97 ms | 0.4830 ms | 1.0903 ms |  1.00 |
| Utf8Parser_Sum |  spanslice | 19.78 ms | 0.3494 ms | 0.3269 ms |  0.88 |
|                |            |          |           |           |       |
|    CountSpaces | 3.0-master | 12.92 ms | 0.2084 ms | 0.1848 ms |  1.00 |
|    CountSpaces |  spanslice | 12.57 ms | 0.1028 ms | 0.0911 ms |  0.97 |

```cs
private readonly byte[] _bytes = Encoding.UTF8.GetBytes("1 22 333 4444 55555 666666 7777777 1 22 333 4444 55555 666666 7777777 1 22 333 4444 55555 666666 7777777 1 22 333 4444 55555 666666 7777777 ");

[Benchmark]
[MethodImpl(MethodImplOptions.AggressiveOptimization)]
public int Utf8Parser_Sum()
{
    byte[] bytes = _bytes;
    _ = bytes.Length; // allow JIT to determine non-null
    int sum = default;

    for (int i = NUM_ITERS; i > 0; i--)
    {
        ReadOnlySpan<byte> copy = bytes;
        while (!copy.IsEmpty)
        {
            Utf8Parser.TryParse(copy, out int value, out int bytesConsumed, 'N');
            copy = copy.Slice(bytesConsumed);
            copy = copy.Slice(1);
            sum += value;
        }
    }

    return sum;
}

[Benchmark]
[MethodImpl(MethodImplOptions.AggressiveOptimization)]
public int CountSpaces()
{
    byte[] bytes = _bytes;
    _ = bytes.Length; // allow JIT to determine non-null
    int count = default;

    for (int i = NUM_ITERS; i > 0; i--)
    {
        ReadOnlySpan<byte> copy = bytes;
        while (!copy.IsEmpty)
        {
            int index = copy.IndexOf((byte)' ');
            if (index < 0) { break; }

            count++;
            copy = copy.Slice(index + 1);
        }
    }

    return count;
}
```

I didn't change the other overloads of `Slice` because I couldn't create a benchmark that clearly demonstrated a perf win.