@GrabYourPitchforks 2019-04-02
The managed changes are very straightforward - use unsigned instead of signed integers when performing byref arithmetic. A shift-F12 on `Unsafe.Add<T>(ref T, int)` will quickly find interesting call sites. Somebody from the JIT side should speak to the required changes there since I didn't see any immediately obvious way to address it.