@jkotas 2019-04-03
I have looked at the asm delta for `Utf8Parser_Sum` before and after your change. It changes `movsxd  rax,eax` into `mov eax,eax`. Is it what you see as well? (I does not make any perf difference on my machine.)

```
mov     rcx,qword ptr [rsi+8]
test    rcx,rcx
jne     test!My.Work()+0x3f (00007ff7`ede3226f)
xor     ebp,ebp
xor     r14d,r14d
jmp     test!My.Work()+0x47 (00007ff7`ede32277)
lea     rbp,[rcx+10h]
mov     r14d,dword ptr [rcx+8]
test    r14d,r14d
jbe     test!My.Work()+0x99 (00007ff7`ede322c9)
lea     rcx,[rsp+20h]
mov     qword ptr [rcx],rbp
mov     dword ptr [rcx+8],r14d
lea     rcx,[rsp+20h]
lea     rdx,[rsp+38h]
lea     r8,[rsp+30h]
mov     r9d,4Eh
call    System.Buffers.Text.Utf8Parser.TryParse(System.ReadOnlySpan`1<Byte>, Int32 ByRef, Int32 ByRef, Char) (00007ff7`ede31f08)
mov     eax,dword ptr [rsp+30h]
cmp     eax,r14d
ja      test!My.Work()+0xac (00007ff7`ede322dc)
sub     r14d,eax
movsxd  rax,eax // this turns into mov eax,eax with your change
add     rbp,rax
cmp     r14d,1
jb      test!My.Work()+0xb2 (00007ff7`ede322e2)
dec     r14d
inc     rbp
add     edi,dword ptr [rsp+38h]
test    r14d,r14d
ja      test!My.Work()+0x4c (00007ff7`ede3227c)
dec     ebx
test    ebx,ebx
jg      test!My.Work()+0x2f (00007ff7`ede3225f)
```