@GrabYourPitchforks 2019-04-03
@jkotas - I also hacked the JIT to eliminate the `mov` entirely. Basically, I "fixed" https://github.com/dotnet/coreclr/issues/23666 and reran the benchmarks to get the numbers you see here. This PR doesn't have that hack (because it was nowhere near production-quality).