@AndyAyersMS 2019-04-03
I don't think there is an easy fix in the jit. There are a bunch of inter-related issues we should tackle in a coordinated fashion:
* Allow cast containment (#12676, responding to #12595, waiting on #20126)
* Don't sign extend for array.len, span.len, and known in-bounds indices when implicit zero extension is available for a 32 bit load (as it is on x64 and arm64). One prototype here: https://github.com/dotnet/coreclr/issues/23127#issuecomment-471040829
* Don't explicitly zero extend values that are already implicitly zero extended (in the spirit of #22454 but done earlier when we have better producer-consumer connections).
* IV widening for 64 bit targets (#9179, #21553)